# api/control-v1.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AWG-Forge node control API",
    "version": "0.1.0-draft",
    "description": "Internal controller/node contract. Enrollment claim, status and presence are implemented only on the explicitly enabled loopback listener. Other operations remain proposed; this draft is not a public operator API. Enrollment uses pinned server TLS and one-time credentials; presence requires registry-authorized mutual TLS.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://controller.example",
      "description": "Dedicated AWG-Forge control listener"
    }
  ],
  "security": [
    {
      "nodeMTLS": []
    }
  ],
...
  "paths": {
    "/control/v1/enrollments/{invitation_id}/claim": {
      "post": {
        "operationId": "claimEnrollmentInvitation",
        "summary": "Claim a one-time enrollment invitation",
        "description": "The node verifies the controller CA/SPKI pin before sending this request. The invitation secret is supplied as a bearer value and must never appear in the URL, command arguments, environment, logs, or durable plaintext storage.",
        "tags": ["Enrollment"],
        "security": [
          {
            "enrollmentSecret": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/InvitationID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollmentClaimRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The CSR is bound to the invitation and awaits explicit controller approval.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentClaimAccepted"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Problem" },
          "401": { "$ref": "#/components/responses/Problem" },
          "404": { "$ref": "#/components/responses/Problem" },
          "409": { "$ref": "#/components/responses/Problem" },
          "410": { "$ref": "#/components/responses/Problem" },
          "413": { "$ref": "#/components/responses/Problem" },
          "429": { "$ref": "#/components/responses/RateLimitedProblem" },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/control/v1/enrollments/{enrollment_id}": {
      "get": {
        "operationId": "getEnrollmentStatus",
        "summary": "Read enrollment approval status",
        "description": "An exact CSR retry returns the same claim capability until expiry; a competing CSR returns 409. The claim token is returned only by the successful claim response and is stored as a digest by the controller. An approved response contains the node certificate but never a private key. Claim, status and presence are implemented for the loopback checkpoint; the other operations remain draft.",
        "tags": ["Enrollment"],
        "security": [
          {
            "claimToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/EnrollmentID"
          }
        ],
        "responses": {
          "200": {
            "description": "Current enrollment state.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestID"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentStatus"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Problem" },
          "404": { "$ref": "#/components/responses/Problem" },
          "410": { "$ref": "#/components/responses/Problem" },
          "429": { "$ref": "#/components/responses/RateLimitedProblem" },
          "default": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "nodeMTLS": {
        "type": "mutualTLS",
        "description": "AWG-Forge-managed node certificate verified by the dedicated control listener and checked against active node and binding state."
      },
      "enrollmentSecret": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque one-time invitation secret",
        "description": "High-entropy single-use enrollment secret entered interactively and accepted only with a verified controller CA/SPKI pin."
      },
      "claimToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque enrollment claim token",
        "description": "Short-lived token returned after the controller atomically binds an invitation to one CSR. Only its digest is stored."
      }
    },
    "parameters": {
      "InvitationID": {
        "name": "invitation_id",
        "in": "path",
        "required": true,
        "description": "Non-secret identifier included in the controller-generated command.",
        "schema": { "$ref": "#/components/schemas/UUID" }
      },
      "EnrollmentID": {
        "name": "enrollment_id",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/UUID" }
      },
...
    },
    "headers": {
      "NoStore": {
        "description": "Control responses are never cacheable.",
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      },
      "RequestID": {
        "description": "Safe correlation identifier for logs and audit events.",
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        }
      },
      "RetryAfter": {
        "description": "Seconds before retrying a rate-limited request.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 3600
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "Safe RFC 9457 problem response without internal error details.",
        "headers": {
          "Cache-Control": { "$ref": "#/components/headers/NoStore" },
          "X-Request-ID": { "$ref": "#/components/headers/RequestID" }
        },
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "RateLimitedProblem": {
        "description": "The request exceeded a bounded enrollment, node, or global limit.",
        "headers": {
          "Cache-Control": { "$ref": "#/components/headers/NoStore" },
          "X-Request-ID": { "$ref": "#/components/headers/RequestID" },
          "Retry-After": { "$ref": "#/components/headers/RetryAfter" }
        },
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      }
    },
    "schemas": {
      "UUID": {
        "type": "string",
        "format": "uuid"
      },
      "BootID": {
        "allOf": [
          { "$ref": "#/components/schemas/UUID" }
        ],
        "description": "Random identity generated on every application process start."
      },
...
      "Capabilities": {
        "type": "array",
        "minItems": 1,
        "maxItems": 128,
        "uniqueItems": true,
        "items": {
          "type": "string",
          "pattern": "^[a-z][a-z0-9]*(?:[._-][a-z0-9]+)*$",
          "maxLength": 64
        }
      },
      "EnrollmentClaimRequest": {
        "type": "object",
        "required": ["requested_name", "csr_pem", "boot_id", "application_version", "contract_versions", "capabilities"],
        "properties": {
          "requested_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "csr_pem": {
            "type": "string",
            "minLength": 128,
            "maxLength": 8192,
            "writeOnly": true,
            "description": "PEM PKCS#10 CSR. The corresponding private key never leaves the node."
          },
          "boot_id": { "$ref": "#/components/schemas/BootID" },
          "application_version": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "contract_versions": {
            "type": "array",
            "minItems": 1,
            "maxItems": 4,
            "uniqueItems": true,
            "items": {
              "type": "integer",
              "minimum": 1,
              "maximum": 255
            }
          },
          "capabilities": { "$ref": "#/components/schemas/Capabilities" }
        },
        "additionalProperties": false
      },
      "EnrollmentClaimAccepted": {
        "type": "object",
        "required": ["enrollment_id", "claim_token", "verification_code", "expires_at", "poll_after_seconds"],
        "properties": {
          "enrollment_id": { "$ref": "#/components/schemas/UUID" },
          "claim_token": {
            "type": "string",
            "minLength": 43,
            "maxLength": 256,
            "readOnly": true,
            "description": "Short-lived opaque token. The controller stores only a keyed digest."
          },
          "verification_code": {
            "type": "string",
            "pattern": "^[A-Z0-9]{4}-[A-Z0-9]{4}$"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "poll_after_seconds": {
            "type": "integer",
            "minimum": 1,
            "maximum": 30
          }
        },
        "additionalProperties": false
      },
      "EnrollmentStatus": {
        "type": "object",
        "oneOf": [
          { "$ref": "#/components/schemas/EnrollmentPending" },
          { "$ref": "#/components/schemas/EnrollmentApproved" },
          { "$ref": "#/components/schemas/EnrollmentRejected" },
          { "$ref": "#/components/schemas/EnrollmentExpired" }
        ],
        "discriminator": {
          "propertyName": "status"
        }
      },
      "EnrollmentPending": {
        "type": "object",
        "required": ["status", "verification_code", "expires_at", "poll_after_seconds"],
        "properties": {
          "status": { "type": "string", "const": "pending" },
          "verification_code": { "type": "string", "pattern": "^[A-Z0-9]{4}-[A-Z0-9]{4}$" },
          "expires_at": { "type": "string", "format": "date-time" },
          "poll_after_seconds": { "type": "integer", "minimum": 1, "maximum": 30 }
        },
        "additionalProperties": false
      },
      "EnrollmentApproved": {
        "type": "object",
        "required": ["status", "node_id", "controller_id", "binding_epoch", "certificate", "contract_version"],
        "properties": {
          "status": { "type": "string", "const": "approved" },
          "node_id": { "$ref": "#/components/schemas/UUID" },
          "controller_id": { "$ref": "#/components/schemas/UUID" },
          "binding_epoch": { "type": "integer", "minimum": 1 },
          "certificate": { "$ref": "#/components/schemas/IssuedCertificate" },
          "contract_version": { "type": "integer", "const": 1 }
        },
        "additionalProperties": false
      },
      "EnrollmentRejected": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "const": "rejected" }
        },
        "additionalProperties": false
      },
      "EnrollmentExpired": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "const": "expired" }
        },
        "additionalProperties": false
      },
...
      "IssuedCertificate": {
        "type": "object",
        "required": ["certificate_pem", "chain_pem", "serial", "not_before", "not_after"],
        "properties": {
          "certificate_pem": { "type": "string", "minLength": 128, "maxLength": 16384, "readOnly": true },
          "chain_pem": { "type": "string", "minLength": 128, "maxLength": 32768, "readOnly": true },
          "serial": { "type": "string", "minLength": 1, "maxLength": 128, "readOnly": true },
          "not_before": { "type": "string", "format": "date-time", "readOnly": true },
          "not_after": { "type": "string", "format": "date-time", "readOnly": true }
        },
        "additionalProperties": false
      },
      "Problem": {
        "type": "object",
        "required": ["type", "title", "status", "code"],
        "properties": {
          "type": { "type": "string", "format": "uri", "maxLength": 512 },
          "title": { "type": "string", "minLength": 1, "maxLength": 128 },
          "status": { "type": "integer", "minimum": 400, "maximum": 599 },
          "detail": { "type": "string", "maxLength": 1024, "description": "Safe actionable detail without stack traces, commands, paths, or secret material." },
          "instance": { "type": "string", "format": "uri-reference", "maxLength": 512 },
          "code": { "type": "string", "pattern": "^[a-z0-9_]+$", "maxLength": 64 }
        },
        "additionalProperties": false
      }
    }
  }
}